GameSubtitleOCR/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "game-subtitle-ocr"
version = "0.1.0"
description = "Extract hardcoded Chinese subtitles from game videos into SRT with PaddleOCR."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "numpy>=1.26.0",
    "opencv-python>=4.10.0.84",
    "paddleocr==3.3.3",
    "tqdm>=4.66.0",
]

[project.scripts]
game-subtitle-ocr = "game_subtitle_ocr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
